Cálculo de double/float en operaciones.

+++ AplicacionInteresante/Program.cs
@@ -7,9 +7,7 @@ namespace AplicacionInteresante
         //Necesitamos:
         static int[] memoria = new int[memoria.Length];
         static int indice = 0;
-        static int linea = 0;//2011
-
-        //string hex = linea.ToString("X");
+        static double linea = 0;//2011 //  de int a double R
 
         static void Main(string[] args)
         {
